Fix some truncation warnings.

+++ Sources/Plasma/CoreLib/hsStream.h
@@ -80,7 +80,7 @@ public:
     virtual void      CopyToMem(void* mem);
     virtual bool      IsCompressed() { return false; }
 
-    uint32_t        WriteString(const ST::string & string) { return Write(string.size(), string.c_str()); }
+    uint32_t        WriteString(const ST::string & string) { return Write((uint32_t)string.size(), string.c_str()); }
 
     template        <typename... _Args>
     uint32_t        WriteFmt(const char * fmt, _Args&&... args) { return WriteString(ST::format(fmt, std::forward<_Args>(args)...)); }
